Moved package to src
Placing the modeci_mdf package under src/ directory helps prevent
accidentally testing against the local directory pacakge versus
the installed package.

Devolopers should install the pacakge with -e to get an editable
install. For example:

pip install -e .[dev]
