+++ setup.cfg
@@ -42,8 +42,12 @@ install_requires =
     graphviz
 python_requires = >=3.6
 include_package_data = True
+package_dir =
+    =src
+
 
 [options.packages.find]
+where = src
 exclude =
     tests
     examples
@@ -78,7 +82,7 @@ per-file-ignores =
 [mypy]
 warn_unused_configs = True
 python_version = 3.6
-files = modeci_mdf
+files = src
 disallow_any_generics = True
 disallow_subclassing_any = True
 disallow_untyped_calls = True
@@ -92,3 +96,18 @@ warn_unused_ignores = True
 warn_return_any = True
 no_implicit_reexport = True
 strict_equality = True
+
+[check-manifest]
+ignore =
+    .all-contributorsrc
+    .pre-commit-config.yaml
+    .readthedocs.yml
+    examples/**
+    notebooks/**
+    docs/**
+    scripts/**
+    dev-requirements.txt
+    CONTRIBUTING.md
+    *.html
+    src/modeci_mdf/version.py
+    tests/.pytest_cache/**
