+++ src/modeci_mdf/export/onnx.py
@@ -50,7 +50,7 @@ def mdf_to_onnx(graph, save_to=None):
 
 if __name__ == "__main__":
 
-    from modeci_mdf.utils import load_mdf_json, print_summary
+    from modeci_mdf import load_mdf_json, print_summary
 
     example = '../../examples/Simple.json'
     verbose = True
